Solved AoC2021 Day 10 Part 2

+++ aoc2021/aoc2021-kotlin/src/test/kotlin/de/havox_design/aoc2021/day10/Day10Test.kt
@@ -15,7 +15,7 @@ class Day10Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart2")
-    fun testProcessPart2(filename: String, expectedResult: Int) =
+    fun testProcessPart2(filename: String, expectedResult: Long) =
         SyntaxScoring(filename).processPart2().shouldBe(expectedResult)
 
     companion object {
@@ -28,7 +28,7 @@ class Day10Test {
         @JvmStatic
         private fun getDataForTestProcessPart2(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2021/day10/day10sample.txt", 288957)
+                Arguments.of("de/havox_design/aoc2021/day10/day10sample.txt", 288957L)
             )
     }
 }
